Issue 1097 - Add a view with a list of students for a course (#1157)
Resolve issue #1097
Add a view of all students enrolled and in queue for a course and each type of groups. The view is available with a "Lista" button on the course page in next to the course name, and each group section.

The view is quite similar to the one for a group. It contains a list of enrolled and queued students, a number of enrolled and queued students, the ability to send an email (hidden addresses or not) and to download a csv list.

We define that a student is enrolled in a course, if he is enrolled in any group in the course. A student is in a queue for a course, if he is in a queue for any group and is not enrolled in any group.